POC-2021: Move useGetDiograph to features/diograph/useGetDiograph.js (from Root.js)
