+++ src/react/Root.js
@@ -1,25 +1,10 @@
-import React, { useEffect } from 'react'
-import { useDispatchActions, useStore } from './store'
+import React from 'react'
+import { useStore } from './store'
 
-import { setFocus } from './features/navigation/actions'
-import { addDiograph } from './features/diograph/actions'
-
-import { invokeChannel } from './client/client'
-import { channels } from '../shared/constants'
+import { useGetDiograph } from './features/diograph/useGetDiograph'
 
 import Browser from './Browser'
 import Welcome from './features/welcome/Welcome'
-
-export const useGetDiograph = () => {
-  const { dispatch } = useDispatchActions()
-
-  useEffect(() => {
-    invokeChannel(channels.GET_DIOGRAPH).then(({ diograph, rootId, folderLocation }) => {
-      dispatch(addDiograph(diograph, rootId, folderLocation))
-      dispatch(setFocus({ id: rootId }))
-    })
-  }, [dispatch])
-}
 
 const Root = () => {
   const [{ folderLocation }] = useStore((state) => state.diograph)
